`Worker`: Move more RTP related classes to `RTC/RTP/` folder
Just that.
