+++ worker/include/RTC/RTP/Codecs/H264.hpp
@@ -70,8 +70,8 @@ namespace RTC
 						this->payloadDescriptor->Dump(indentation);
 					}
 					bool Process(
-					  Codecs::EncodingContext* encodingContext, RTC::RTP::Packet* packet, bool& marker) override;
-					void RtpPacketChanged(RTC::RTP::Packet* packet) override {};
+					  Codecs::EncodingContext* encodingContext, RTP::Packet* packet, bool& marker) override;
+					void RtpPacketChanged(RTP::Packet* packet) override {};
 					std::unique_ptr<Codecs::PayloadDescriptor::Encoder> GetEncoder() const override
 					{
 						return nullptr;
