+++ worker/include/RTC/SimulcastConsumer.hpp
@@ -10,7 +10,7 @@
 
 namespace RTC
 {
-	class SimulcastConsumer : public RTC::Consumer, public RTC::RtpStreamSend::Listener
+	class SimulcastConsumer : public RTC::Consumer, public RTC::RTP::RtpStreamSend::Listener
 	{
 	public:
 		SimulcastConsumer(
@@ -45,7 +45,7 @@ namespace RTC
 				std::any_of(
 					this->producerRtpStreams.begin(),
 					this->producerRtpStreams.end(),
-					[](const RTC::RtpStreamRecv* rtpStream)
+					[](const RTC::RTP::RtpStreamRecv* rtpStream)
 					{
 						// If there is no RTP inactivity check do not consider the stream
 						// inactive despite it has score 0.
@@ -55,18 +55,18 @@ namespace RTC
 			);
 			// clang-format on
 		}
-		void ProducerRtpStream(RTC::RtpStreamRecv* rtpStream, uint32_t mappedSsrc) override;
-		void ProducerNewRtpStream(RTC::RtpStreamRecv* rtpStream, uint32_t mappedSsrc) override;
+		void ProducerRtpStream(RTC::RTP::RtpStreamRecv* rtpStream, uint32_t mappedSsrc) override;
+		void ProducerNewRtpStream(RTC::RTP::RtpStreamRecv* rtpStream, uint32_t mappedSsrc) override;
 		void ProducerRtpStreamScore(
-		  RTC::RtpStreamRecv* rtpStream, uint8_t score, uint8_t previousScore) override;
-		void ProducerRtcpSenderReport(RTC::RtpStreamRecv* rtpStream, bool first) override;
+		  RTC::RTP::RtpStreamRecv* rtpStream, uint8_t score, uint8_t previousScore) override;
+		void ProducerRtcpSenderReport(RTC::RTP::RtpStreamRecv* rtpStream, bool first) override;
 		uint8_t GetBitratePriority() const override;
 		uint32_t IncreaseLayer(uint32_t bitrate, bool considerLoss) override;
 		void ApplyLayers() override;
 		uint32_t GetDesiredBitrate() const override;
 		void SendRtpPacket(RTC::RTP::Packet* packet, RTC::RTP::SharedPacket& sharedPacket) override;
 		bool GetRtcp(RTC::RTCP::CompoundPacket* packet, uint64_t nowMs) override;
-		const std::vector<RTC::RtpStreamSend*>& GetRtpStreams() const override
+		const std::vector<RTC::RTP::RtpStreamSend*>& GetRtpStreams() const override
 		{
 			return this->rtpStreams;
 		}
@@ -99,22 +99,23 @@ namespace RTC
 		void StorePacketInTargetLayerRetransmissionBuffer(
 		  RTC::RTP::Packet* packet, RTC::RTP::SharedPacket& sharedPacket);
 		void EmitLayersChange() const;
-		RTC::RtpStreamRecv* GetProducerCurrentRtpStream() const;
-		RTC::RtpStreamRecv* GetProducerTargetRtpStream() const;
-		RTC::RtpStreamRecv* GetProducerTsReferenceRtpStream() const;
+		RTC::RTP::RtpStreamRecv* GetProducerCurrentRtpStream() const;
+		RTC::RTP::RtpStreamRecv* GetProducerTargetRtpStream() const;
+		RTC::RTP::RtpStreamRecv* GetProducerTsReferenceRtpStream() const;
 
 		/* Pure virtual methods inherited from RtpStreamSend::Listener. */
 	public:
-		void OnRtpStreamScore(RTC::RtpStream* rtpStream, uint8_t score, uint8_t previousScore) override;
-		void OnRtpStreamRetransmitRtpPacket(RTC::RtpStreamSend* rtpStream, RTC::RTP::Packet* packet) override;
+		void OnRtpStreamScore(RTC::RTP::RtpStream* rtpStream, uint8_t score, uint8_t previousScore) override;
+		void OnRtpStreamRetransmitRtpPacket(
+		  RTC::RTP::RtpStreamSend* rtpStream, RTC::RTP::Packet* packet) override;
 
 	private:
 		// Allocated by this.
-		RTC::RtpStreamSend* rtpStream{ nullptr };
+		RTC::RTP::RtpStreamSend* rtpStream{ nullptr };
 		// Others.
 		absl::flat_hash_map<uint32_t, int16_t> mapMappedSsrcSpatialLayer;
-		std::vector<RTC::RtpStreamSend*> rtpStreams;
-		std::vector<RTC::RtpStreamRecv*> producerRtpStreams; // Indexed by spatial layer.
+		std::vector<RTC::RTP::RtpStreamSend*> rtpStreams;
+		std::vector<RTC::RTP::RtpStreamRecv*> producerRtpStreams; // Indexed by spatial layer.
 		bool syncRequired{ false };
 		int16_t spatialLayerToSync{ -1 };
 		bool lastSentPacketHasMarker{ false };
