+++ worker/meson.build
@@ -120,11 +120,6 @@ common_sources = [
   'src/RTC/RtcLogger.cpp',
   'src/RTC/RtpListener.cpp',
   'src/RTC/RtpObserver.cpp',
-  'src/RTC/RtpRetransmissionBuffer.cpp',
-  'src/RTC/RtpStream.cpp',
-  'src/RTC/RtpStreamRecv.cpp',
-  'src/RTC/RtpStreamSend.cpp',
-  'src/RTC/RtxStream.cpp',
   'src/RTC/SctpAssociation.cpp',
   'src/RTC/SctpListener.cpp',
   'src/RTC/SenderBandwidthEstimator.cpp',
@@ -161,6 +156,11 @@ common_sources = [
   'src/RTC/ICE/StunPacket.cpp',
   'src/RTC/RTP/Packet.cpp',
   'src/RTC/RTP/ProbationGenerator.cpp',
+  'src/RTC/RTP/RetransmissionBuffer.cpp',
+  'src/RTC/RTP/RtpStream.cpp',
+  'src/RTC/RTP/RtpStreamRecv.cpp',
+  'src/RTC/RTP/RtpStreamSend.cpp',
+  'src/RTC/RTP/RtxStream.cpp',
   'src/RTC/RTP/SharedPacket.cpp',
   'src/RTC/RTP/Codecs/AV1.cpp',
   'src/RTC/RTP/Codecs/H264.cpp',
@@ -409,9 +409,6 @@ test_sources = [
   'test/src/RTC/TestNackGenerator.cpp',
   'test/src/RTC/TestRateCalculator.cpp',
   'test/src/RTC/TestRtpEncodingParameters.cpp',
-  'test/src/RTC/TestRtpRetransmissionBuffer.cpp',
-  'test/src/RTC/TestRtpStreamSend.cpp',
-  'test/src/RTC/TestRtpStreamRecv.cpp',
   'test/src/RTC/TestSeqManager.cpp',
   'test/src/RTC/TestSimpleConsumer.cpp',
   'test/src/RTC/TestTransportCongestionControlServer.cpp',
@@ -422,6 +419,9 @@ test_sources = [
   'test/src/RTC/RTP/rtpCommon.cpp',
   'test/src/RTC/RTP/TestPacket.cpp',
   'test/src/RTC/RTP/TestProbationGenerator.cpp',
+  'test/src/RTC/RTP/TestRetransmissionBuffer.cpp',
+  'test/src/RTC/RTP/TestRtpStreamSend.cpp',
+  'test/src/RTC/RTP/TestRtpStreamRecv.cpp',
   'test/src/RTC/RTP/TestSharedPacket.cpp',
   'test/src/RTC/RTP/Codecs/TestH264.cpp',
   'test/src/RTC/RTP/Codecs/TestVP8.cpp',
@@ -632,13 +632,13 @@ fuzzer_sources = common_sources + [
   'fuzzer/src/FuzzerUtils.cpp',
   'fuzzer/src/RTC/FuzzerDtlsTransport.cpp',
   'fuzzer/src/RTC/FuzzerRateCalculator.cpp',
-  'fuzzer/src/RTC/FuzzerRtpRetransmissionBuffer.cpp',
-  'fuzzer/src/RTC/FuzzerRtpStreamSend.cpp',
   'fuzzer/src/RTC/FuzzerSeqManager.cpp',
   'fuzzer/src/RTC/FuzzerTrendCalculator.cpp',
   'fuzzer/src/RTC/ICE/FuzzerStunPacket.cpp',
   'fuzzer/src/RTC/RTP/FuzzerPacket.cpp',
   'fuzzer/src/RTC/RTP/FuzzerProbationGenerator.cpp',
+  'fuzzer/src/RTC/RTP/FuzzerRetransmissionBuffer.cpp',
+  'fuzzer/src/RTC/RTP/FuzzerRtpStreamSend.cpp',
   'fuzzer/src/RTC/RTP/Codecs/FuzzerOpus.cpp',
   'fuzzer/src/RTC/RTP/Codecs/FuzzerAV1.cpp',
   'fuzzer/src/RTC/RTP/Codecs/FuzzerH264.cpp',
