+++ worker/src/RTC/Router.cpp
@@ -606,7 +606,7 @@ namespace RTC
 	inline void Router::OnTransportProducerNewRtpStream(
 	  RTC::Transport* /*transport*/,
 	  RTC::Producer* producer,
-	  RTC::RtpStreamRecv* rtpStream,
+	  RTC::RTP::RtpStreamRecv* rtpStream,
 	  uint32_t mappedSsrc)
 	{
 		MS_TRACE();
@@ -622,7 +622,7 @@ namespace RTC
 	inline void Router::OnTransportProducerRtpStreamScore(
 	  RTC::Transport* /*transport*/,
 	  RTC::Producer* producer,
-	  RTC::RtpStreamRecv* rtpStream,
+	  RTC::RTP::RtpStreamRecv* rtpStream,
 	  uint8_t score,
 	  uint8_t previousScore)
 	{
@@ -637,7 +637,7 @@ namespace RTC
 	}
 
 	inline void Router::OnTransportProducerRtcpSenderReport(
-	  RTC::Transport* /*transport*/, RTC::Producer* producer, RTC::RtpStreamRecv* rtpStream, bool first)
+	  RTC::Transport* /*transport*/, RTC::Producer* producer, RTC::RTP::RtpStreamRecv* rtpStream, bool first)
 	{
 		MS_TRACE();
