+++ worker/src/RTC/Transport.cpp
@@ -2422,7 +2422,7 @@ namespace RTC
 	}
 
 	inline void Transport::OnProducerNewRtpStream(
-	  RTC::Producer* producer, RTC::RtpStreamRecv* rtpStream, uint32_t mappedSsrc)
+	  RTC::Producer* producer, RTC::RTP::RtpStreamRecv* rtpStream, uint32_t mappedSsrc)
 	{
 		MS_TRACE();
 
@@ -2430,7 +2430,7 @@ namespace RTC
 	}
 
 	inline void Transport::OnProducerRtpStreamScore(
-	  RTC::Producer* producer, RTC::RtpStreamRecv* rtpStream, uint8_t score, uint8_t previousScore)
+	  RTC::Producer* producer, RTC::RTP::RtpStreamRecv* rtpStream, uint8_t score, uint8_t previousScore)
 	{
 		MS_TRACE();
 
@@ -2438,7 +2438,7 @@ namespace RTC
 	}
 
 	inline void Transport::OnProducerRtcpSenderReport(
-	  RTC::Producer* producer, RTC::RtpStreamRecv* rtpStream, bool first)
+	  RTC::Producer* producer, RTC::RTP::RtpStreamRecv* rtpStream, bool first)
 	{
 		MS_TRACE();
